+++ compiler/src/jdk.graal.compiler.test/src/jdk/graal/compiler/util/args/test/OptionValueTest.java
@@ -22,18 +22,19 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package org.graalvm.igvutil.test.args;
+package jdk.graal.compiler.util.args.test;
 
 import java.util.List;
 
-import org.graalvm.igvutil.args.BooleanValue;
-import org.graalvm.igvutil.args.Flag;
-import org.graalvm.igvutil.args.IntegerValue;
-import org.graalvm.igvutil.args.InvalidArgumentException;
-import org.graalvm.igvutil.args.OptionValue;
-import org.graalvm.igvutil.args.StringValue;
 import org.junit.Assert;
 import org.junit.Test;
+
+import jdk.graal.compiler.util.args.BooleanValue;
+import jdk.graal.compiler.util.args.Flag;
+import jdk.graal.compiler.util.args.IntegerValue;
+import jdk.graal.compiler.util.args.InvalidArgumentException;
+import jdk.graal.compiler.util.args.OptionValue;
+import jdk.graal.compiler.util.args.StringValue;
 
 public class OptionValueTest {
