+++ compiler/src/jdk.graal.compiler/src/jdk/graal/compiler/util/args/Command.java
@@ -22,7 +22,7 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package org.graalvm.igvutil.args;
+package jdk.graal.compiler.util.args;
 
 import java.io.PrintWriter;
 import java.util.ArrayList;
