+++ compiler/src/org.graalvm.igvutil/src/org/graalvm/igvutil/IgvUtility.java
@@ -34,16 +34,15 @@ import java.util.List;
 import java.util.Set;
 import java.util.zip.GZIPInputStream;
 
-import org.graalvm.igvutil.args.Command;
-import org.graalvm.igvutil.args.CommandGroup;
-import org.graalvm.igvutil.args.OptionValue;
-import org.graalvm.igvutil.args.Program;
-import org.graalvm.igvutil.args.StringValue;
-
 import jdk.graal.compiler.graphio.parsing.BinaryReader;
 import jdk.graal.compiler.graphio.parsing.ModelBuilder;
 import jdk.graal.compiler.graphio.parsing.StreamSource;
 import jdk.graal.compiler.graphio.parsing.model.GraphDocument;
+import jdk.graal.compiler.util.args.Command;
+import jdk.graal.compiler.util.args.CommandGroup;
+import jdk.graal.compiler.util.args.OptionValue;
+import jdk.graal.compiler.util.args.Program;
+import jdk.graal.compiler.util.args.StringValue;
 import jdk.graal.compiler.util.json.JsonBuilder;
 import jdk.graal.compiler.util.json.JsonPrettyWriter;
 import jdk.graal.compiler.util.json.JsonWriter;
